feat(chat,presence,notifications): active-status reporter, call notifications, chat UX polish
- Add ActiveStatusReporter + UpdateActiveStatusResponse to bump
  users.datetime via update_active_status (4-min client throttle,
  reset on logout)
- Wire ApiManager + BaseApplication + MainActivity for presence pings
- Add CallNotifications utility for rich messaging-style call alerts
  with avatars, shortcuts and action receivers
- Rework chat experience: ChatActivityInHouse, ChatAdapter,
  ChatListAdapter, ChatAudioPlayer, ChatHistoryMemoryCache,
  ChatNotifications/Store, ImageCompressor, ActiveChatTracker,
  PinnedChatsPrefsHelper
- Polish chat layouts (sent/received message variants, audio/image
  bubbles, attach sheet, reply quote, conversation row, chat list,
  popup menu) and supporting drawables (reaction circle, input bg,
  green circle, ic_arrow_down, ic_image_broken)
- Fragments: FriendsTabFragment, FavouriteFragment, HomeFragment,
  CreatorChatFragment refinements
- Update OneSignalNotificationServiceExtension and FCM service
  for cross-user routing and chat header fallbacks
- Misc: ChatMessage(Api) models, ProfileViewModel, NewLoginActivity,
  UserProfileDetailActivity, BottomSheetLogout, SocketManager,
  AndroidManifest, colors/strings, docs/info
- Ignore .claude/ local agent state

Made-with: Cursor

+++ app/src/main/java/com/gmwapp/hima/models/ChatMessageApi.kt
@@ -2,8 +2,10 @@ package com.gmwapp.hima.models
 
 import com.google.gson.annotations.SerializedName
 
+// T4: id widened to Long to keep parity with the socket payload and avoid
+// truncation when the backend emits snowflake-style ids beyond Int.MAX_VALUE.
 data class ChatMessageApi(
-    val id: Int,
+    val id: Long,
     @SerializedName("chat_id")
     val chatId: String,
     val from: String,
